insurance-api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
python-multipart==0.0.6
python-dotenv==1.0.0
fastapi-mcp==0.1.0
langchain-openai==0.0.5
mcp-use==1.3.9
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
